fix 10 pending unit tests moving markup points over forbidden ranges
Changing `before` and `behind` by excluding the equal cases (<=, >=)
does the trick.

In case of equality and markup like XML, the annotation ends right on
the first markup character, which is a case of `*forbidden`.

We micht also have fixed the thing by moving the `*forbidden` testing
in 'merge' before testing 'before' and 'behind'.