replace ShrinkingNode type class by partial evaluation
This allows us to move the configuration for shrinking XML into the
XML-related module DomTypeDefs. We can now pass a partially evaluated
function has been configured somehow to the shrinkedText
function. This offers a cleaner abstraction from a concrete markup
language.

Please note, that I've been trying to add a ShrinkingNodeConfig type
class in order to achieve such an abstraction. But that failed because
which a type class for shrinking nodes and a shrinking configs, it is
impossible to tie a specific node, XML, to a specific config, e.g. for
XML, because of Haskell's all quantification.